# docs/reference/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Keytography.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/vault/entries": {
      "post": {
        "tags": [
          "VaultEndpoints"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVaultEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultEntryDetailResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "VaultEndpoints"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VaultEntryListItemResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/vault/entries/{id}": {
      "get": {
        "tags": [
          "VaultEndpoints"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultEntryDetailResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "VaultEndpoints"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVaultEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultEntryDetailResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "VaultEndpoints"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/vault/entries/{id}/history": {
      "get": {
        "tags": [
          "VaultEndpoints"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VaultEntryHistoryItemResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/vault/entries/{id}/restore": {
      "post": {
        "tags": [
          "VaultEndpoints"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/vault/entries/{id}/permanent": {
      "delete": {
        "tags": [
          "VaultEndpoints"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/vault/users/{userId}/entries": {
      "get": {
        "tags": [
          "VaultEndpoints"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VaultEntryListItemResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Keytography.Api"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/vault/users/{userId}/entries/{id}": {
      "get": {
        "tags": [
          "VaultEndpoints"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultEntryDetailResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Keytography.Api"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Keytography.Api"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateVaultEntryRequest": {
        "required": [
          "title",
          "login",
          "password",
          "additionalFields"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "login": {
            "type": [
              "null",
              "string"
            ]
          },
          "password": {
            "type": "string"
          },
          "additionalFields": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "CriterionEvaluation": {
        "required": [
          "score",
          "passed"
        ],
        "type": "object",
        "properties": {
          "score": {
            "type": "number",
            "format": "double"
          },
          "passed": {
            "type": "boolean"
          }
        }
      },
...
      "UpdateVaultEntryRequest": {
        "required": [
          "title",
          "login",
          "password",
          "additionalFields"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "login": {
            "type": [
              "null",
              "string"
            ]
          },
          "password": {
            "type": "string"
          },
          "additionalFields": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "VaultEntryDetailResponse": {
        "required": [
          "id",
          "title",
          "login",
          "password",
          "additionalFields",
          "createdAt",
          "updatedAt",
          "passwordScore",
          "passwordScoreDetail"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "login": {
            "type": [
              "null",
              "string"
            ]
          },
          "password": {
            "type": "string"
          },
          "additionalFields": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "passwordScore": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "passwordScoreDetail": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "$ref": "#/components/schemas/CriterionEvaluation"
            }
          }
        }
      },
      "VaultEntryHistoryItemResponse": {
        "required": [
          "id",
          "password",
          "changedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "password": {
            "type": "string"
          },
          "changedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "VaultEntryListItemResponse": {
        "required": [
          "id",
          "title",
          "login",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "login": {
            "type": [
              "null",
              "string"
            ]
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
}